requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-cors-headers==3.8.0
django-phonenumber-field==5.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
idna==3.2
phonenumbers==8.12.31
Pillow==8.3.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
